backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.11
langchain_openai==0.0.8
pydantic==2.6.4
python-dotenv==1.0.1
SQLAlchemy==2.0.28
openai==0.27.6
faiss-cpu==1.7.4
python-multipart==0.0.9
PyMuPDF==1.23.26


# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2
